Adds UI functionality
