+++ src/main/java/backend/BinaryOperations.java
@@ -1,0 +1,35 @@
+package backend;
+
+public class BinaryOperations {
+
+
+    /// Returns the value obtained by multiplying b1 and b2
+    public byte Multiply(byte b1, byte b2) {
+        return 0;
+    }
+
+    /// Returns the value obtained by dividing b2 into b1
+    public byte Divide(byte b1, byte b2) {
+        return 0;
+    }
+
+    /// Returns the value obtained by adding b1 to b2
+    public byte Add(byte b1, byte b2) {
+        return 0;
+    }
+
+    /// Returns the value obtained by subtracting b2 from b1
+    public byte Subtract(byte b1, byte b2) {
+        return 0;
+    }
+
+    /// Returns the decimal representation of the passed in number
+    public int ToDecimal(byte number) {
+        return 0;
+    }
+
+    /// Returns the binary representation of the passed in number
+    public byte ToBinary(int number) {
+        return 0;
+    }
+}
