+++ src/main/java/module-info.java
@@ -1,0 +1,6 @@
+module javafx {
+    requires javafx.controls;
+    requires javafx.fxml;
+    opens ui;
+    opens backend;
+}
