+++ src/main/java/ui/Main.java
@@ -1,22 +1,24 @@
 package ui;
 
+
+
 import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.fxml.FXMLLoader;
 import javafx.stage.Stage;
+
+import java.io.IOException;
+import java.util.Objects;
 
 public class Main extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception {
-        System.out.println(getClass().getClassLoader().getResource("sample.fxml"));
-        Parent root = FXMLLoader.load(getClass().getClassLoader().getResource("sample.fxml"));
-        primaryStage.setTitle("Binary Calculator");
-        primaryStage.setScene(new Scene(root, 300, 275));
+        Parent mainScene = FXMLLoader.load(Objects.requireNonNull(getClass().getClassLoader().getResource("sample.fxml")));
+        primaryStage.setScene(new Scene(mainScene));
         primaryStage.show();
     }
-
     public static void main(String[] args) {
         launch(args);
     }
