+++ src/main/resources/sample.fxml
@@ -1,14 +1,73 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import java.lang.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.text.Font?>
 
-<AnchorPane xmlns="http://javafx.com/javafx"
-            xmlns:fx="http://javafx.com/fxml"
-            fx:controller="ui.Main"
-            prefHeight="400.0" prefWidth="600.0">
 
+<AnchorPane prefHeight="531.0" prefWidth="435.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ui.Main">
+   <children>
+      <Label alignment="CENTER" layoutX="91.0" layoutY="14.0" prefHeight="45.0" prefWidth="253.0" text="Binary Calculator">
+         <font>
+            <Font name="System Bold" size="30.0" />
+         </font>
+      </Label>
+      <Button layoutX="91.0" layoutY="102.0" mnemonicParsing="false" prefHeight="75.0" prefWidth="59.0" text="1">
+         <font>
+            <Font size="35.0" />
+         </font>
+      </Button>
+      <Button layoutX="262.0" layoutY="102.0" mnemonicParsing="false" prefHeight="75.0" prefWidth="59.0" text="0">
+         <font>
+            <Font size="35.0" />
+         </font>
+      </Button>
+      <Label layoutX="99.0" layoutY="482.0" prefHeight="30.0" prefWidth="67.0" text="Result:">
+         <font>
+            <Font size="20.0" />
+         </font>
+      </Label>
+      <Label layoutX="175.0" layoutY="471.0" prefHeight="52.0" prefWidth="242.0" text="0" textAlignment="CENTER">
+         <font>
+            <Font size="30.0" />
+         </font>
+      </Label>
+      <Button layoutX="99.0" layoutY="446.0" mnemonicParsing="false" text="Toggle Binary/Decimal" />
+      <Button layoutX="99.0" layoutY="222.0" mnemonicParsing="false" text="+">
+         <font>
+            <Font size="20.0" />
+         </font>
+      </Button>
+      <Button layoutX="190.0" layoutY="222.0" mnemonicParsing="false" prefHeight="44.0" prefWidth="42.0" text="-">
+         <font>
+            <Font size="20.0" />
+         </font>
+      </Button>
+      <Button layoutX="287.0" layoutY="222.0" mnemonicParsing="false" prefHeight="44.0" prefWidth="42.0" text="*">
+         <font>
+            <Font size="20.0" />
+         </font>
+      </Button>
+      <Button layoutX="99.0" layoutY="287.0" mnemonicParsing="false" prefHeight="44.0" prefWidth="42.0" text="/">
+         <font>
+            <Font size="20.0" />
+         </font>
+      </Button>
+      <Button layoutX="190.0" layoutY="287.0" mnemonicParsing="false" prefHeight="44.0" prefWidth="42.0" text="√">
+         <font>
+            <Font size="20.0" />
+         </font>
+      </Button>
+      <Button layoutX="284.0" layoutY="287.0" mnemonicParsing="false" prefHeight="44.0" prefWidth="49.0" text="x²">
+         <font>
+            <Font size="20.0" />
+         </font>
+      </Button>
+      <Button layoutX="99.0" layoutY="342.0" mnemonicParsing="false" prefHeight="54.0" prefWidth="234.0" text="=">
+         <font>
+            <Font size="25.0" />
+         </font>
+      </Button>
+   </children>
 </AnchorPane>
